Extracted the publisher's dirargs to separate file
